Split GUI app lifecycle
